triton/requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile base.in
#
aiohttp==3.8.1
    # via
    #   datasets
    #   fsspec
aiosignal==1.2.0
    # via aiohttp
altgraph==0.17.2
    # via
    #   macholib
    #   pyinstaller
argparse==1.4.0
    # via -r base.in
async-timeout==4.0.2
    # via aiohttp
attrs==21.4.0
    # via aiohttp
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via
    #   aiohttp
    #   requests
click==8.1.2
    # via sacremoses
datasets==2.3.2
    # via -r base.in
dill==0.3.5.1
    # via
    #   datasets
    #   multiprocess
filelock==3.6.0
    # via
    #   huggingface-hub
    #   transformers
flatbuffers==2.0
    # via onnxruntime
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2022.5.0
    # via datasets
grpcio==1.47.0
    # via
    #   -r base.in
    #   grpcio-tools
grpcio-tools==1.47.0
    # via -r base.in
huggingface-hub==0.5.1
    # via
    #   datasets
    #   transformers
idna==3.3
    # via
    #   requests
    #   yarl
joblib==1.1.0
    # via sacremoses
macholib==1.16
    # via pyinstaller
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.13
    # via datasets
numpy==1.22.3
    # via
    #   datasets
    #   onnxruntime
    #   pandas
    #   pyarrow
    #   transformers
onnxruntime==1.11.0
    # via -r base.in
packaging==21.3
    # via
    #   datasets
    #   huggingface-hub
    #   transformers
pandas==1.4.3
    # via datasets
protobuf==3.20.0
    # via
    #   grpcio-tools
    #   onnxruntime
pyarrow==8.0.0
    # via datasets
pyinstaller==5.0
    # via -r base.in
pyinstaller-hooks-contrib==2022.4
    # via pyinstaller
pyparsing==3.0.8
    # via packaging
python-dateutil==2.8.2
    # via pandas
pytz==2022.1
    # via pandas
pyyaml==6.0
    # via
    #   huggingface-hub
    #   transformers
regex==2022.3.15
    # via
    #   sacremoses
    #   transformers
requests==2.27.1
    # via
    #   datasets
    #   fsspec
    #   huggingface-hub
    #   responses
    #   transformers
responses==0.18.0
    # via datasets
sacremoses==0.0.49
    # via transformers
sentencepiece==0.1.96
    # via -r base.in
six==1.16.0
    # via
    #   grpcio
    #   python-dateutil
    #   sacremoses
tokenizers==0.12.1
    # via transformers
torch==1.11.0
    # via -r base.in
tqdm==4.64.0
    # via
    #   datasets
    #   huggingface-hub
    #   sacremoses
    #   transformers
transformers==4.20.0
    # via -r base.in
typing-extensions==4.2.0
    # via
    #   huggingface-hub
    #   torch
urllib3==1.26.9
    # via
    #   requests
    #   responses
xxhash==3.0.0
    # via datasets
yarl==1.7.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
